bug with mortality plots + shp projection correction
Don't know why suddenly the age-adjusted mortality plot (line 258) is
suddenly not working! Also. changes to shp file from CPH_quarter3 to
Quarters2 was made because Quarters3 was projected to match the
projection of this historical raster map - but this projection does not
match the webmaps that I use as a background in this file