Optimize slightly for-loop of full sim
Initialize lists to length of loops (http://goo.gl/9xO3HN) and move
several of the matrix initializations outside of the for-loop. They will
be overwritten in the loop, and the initial states will remain the same
(i.e. the outcome of the previous loop iteration will not affect the
starting conditions of the following loop).