Convert lists to matrices in internal loop
Simulations for each phi now stored as one matrix instead of as a list
of vectors. This reduces one level of nesting within the data structure
and makes the data more similar to the output of the parallel code.

NEXT: Use serial output in LL code to see where things are going
different as compared to the parallel output