Stan model failure
Stan cannot work. There is unobserved data (recovered ppl) that are the
result of a poisson process - so it must be an integer. However we
cannot define an integer as a parameter or transformed parameter value.
According to: goo.gl/Uipqmx only JAGS can do this. Will confirm with
Marina, to make sure there is no work-around that I have not seen