Checked indexing on beta matrix
Inserted a fake fixed beta value to ensure that the indexing from the
JAGS file was consistant with the indexing in the mcmc outputs and beta
matrix results section. I replaces beta[3,2] with a fixed value. This
fixed value showed up in the correct location in the results. No more
need to worry about indexing!!